CPR-905 move getType method
